(Refactor) Refactored ultralytics to the new lib structure.
