+++ icevision/models/ultralytics/yolov5/backbones/__init__.py
@@ -1,0 +1,1 @@
+from icevision.models.ultralytics.yolov5.backbones.backbones import *
