--- icevision/models/ultralytics/yolov5/fastai/__init__.py
@@ -1,2 +1,0 @@
-from icevision.models.ultralytics.yolov5.fastai.callbacks import *
-from icevision.models.ultralytics.yolov5.fastai.learner import *
