--- icevision/models/ultralytics/yolov5/lightning/__init__.py
@@ -1,1 +1,0 @@
-from icevision.models.ultralytics.yolov5.lightning.model_adapter import *
